changed chromosome regex to format correctly. Changed pipeline to work with CNA data
Because
*CNA is a did not work as a single interval
*regex was not working

Changed the executLiftOver() to return a List<String>. Then genomcoordinates are parsed into a flat list of string as 'CHR, sequenceStartPos, sequenceEndPos'. If the point is CNA data there will be six entries into the list 'CHR, sequenceStartPos, sequenceStartPos, CHR sequenceEndPos, sequenceEndPos'. The harmonize this will then be put into a single row for appropriate CNA data. The executeLiftOver will runn .foreach on every single entry passed to it in the genomeCoord. However, for mut data there will only be one entry.